Reintroduced a limited form of LKI, which only saves the most recent role, and only until a new zone changes. This is fine for the demo decks, but breaks down in more complicated scenarios:
Creature is originally green. Make it red. It's ping 1 ability is on the stack. Blink the creature, make it blue, put another ping ability on the stack and then destroy it. The topmost ability should deal damage from a blue source, and the next one from a red source.

Also, since all roles have the same characteristics, the attribute access will never fall through to the lki object in __getattr__, so basically the only attributes that will work are power/toughness.

The way to fix this is basically remove all references to the GameObject in the code, and only use GameRoles. This is a huge undertaking and probably requires a rewrite.