Updates to the UI. Now that the stack is true zone, no longer need hack to hide cards played from hand
Other big change is that CardLibrary no longer caches Card objects, since that was leading to weird behavior where the gui card object was created with a GameObject, but a CardRole was requested (especially in the Zone.get_card_order function). Basically the first time the visual card object was created it was stuck with that gamecard attribute.